feat: port YouTube downloader and GenericYtdlpDownloader to core
Phase 0+1:
- Move GenericYtdlpDownloader to omniget-core/platforms/generic_ytdlp.rs
- Port YouTubeDownloader to omniget-core/platforms/youtube.rs
- Core platform registry exports GenericYtdlpDownloader, YouTubeDownloader, PlatformDownloader
- Tauri layer now re-exports from core instead of owning the code
- All 3 YouTube unit tests pass
- Full workspace compiles (Tauri + CLI)
